Add initial How to Build with Stuart Document
Adds a new document that comprehensively describes how to manually
set up a build environment using Stuart.

Covers the following topics:

1. Pre-requisites - Git, Python, compilers, SDKs, etc.
2. Initial steps - Cloning the repo, set up Python virtual env, etc.
3. Stuart command explanation and corresponding examples
4. Common developer scenarios with examples showing what to do
5. Common questions section with answers

Stuart presents a relatively consistent interface to build most
content in edk2 so this document is meant as a long-term starting
point for new contributors to understand the basics of using Stuart
for various areas they are working in within edk2.

Cc: Sean Brogan <sean.brogan@microsoft.com>
Cc: Michael D Kinney <michael.d.kinney@intel.com>
Cc: Liming Gao <gaoliming@byosoft.com.cn>
Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>