Laszlo's unkempt git guide: improve git-send-email options
(1) Recent git-send-email auto-CC's email addresses from miscellaneous
"Whatever-by:" tags; suppress that logic as well. (What we really want is
to restrict the CC'ing logic to "bodycc" + "cccmd", but there is no way to
state that in a positive sense, so we need to suppress "everything else".)

(2) Embedded CRLFs are best protected with base64
content-transfer-encoding; add an according option. Quoted-printable
encoding tends to wreak havoc for both git-am and mailing list archives on
the web.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>